Create Mortgage Amortization Schedule
An amortization schedule is a complete table of periodic loan payments , showing the amount of principal and the amount of interest that comprise each payment until the loan is paid off at the end of its term.
While each periodic payment is the same amount early in the schedule, the majority of each payment is interest; later in the schedule, the majority of each payment covers the loan's principal. The last line of the schedule shows the borrower’s total interest and principal payments for the entire loan term. See excel file for more details.