Variable Term Mortgage Amortization Schedule
With a Closed Variable Interest Rate Mortgage, when the interest rate changes, the monthly payment amount remains the same. However, the amount that is applied toward interest and principal will change. If the interest rate decreases, more of your payment is applied to the principal. If the interest rate increases, more of the payment will go toward the interest accruing on your mortgage